+++ client/http-client/src/transport.rs
@@ -104,6 +104,8 @@ pub struct HttpTransportClientBuilder<L> {
 	pub(crate) keep_alive_interval: Option<std::time::Duration>,
 	/// KEEP_ALIVE retries
 	pub(crate) keep_alive_retries: Option<u32>,
+	/// TCP connect timeout
+	pub(crate) connect_timeout: Option<std::time::Duration>,
 }
 
 impl Default for HttpTransportClientBuilder<Identity> {
@@ -126,6 +128,7 @@ impl HttpTransportClientBuilder<Identity> {
 			keep_alive_duration: None,
 			keep_alive_interval: None,
 			keep_alive_retries: None,
+			connect_timeout: None,
 		}
 	}
 }
@@ -197,6 +200,7 @@ impl<L> HttpTransportClientBuilder<L> {
 			keep_alive_duration: self.keep_alive_duration,
 			keep_alive_retries: self.keep_alive_retries,
 			keep_alive_interval: self.keep_alive_interval,
+			connect_timeout: self.connect_timeout,
 		}
 	}
 
@@ -220,6 +224,7 @@ impl<L> HttpTransportClientBuilder<L> {
 			keep_alive_duration,
 			keep_alive_interval,
 			keep_alive_retries,
+			connect_timeout,
 		} = self;
 		let mut url = Url::parse(target.as_ref()).map_err(|e| Error::Url(format!("Invalid URL: {e}")))?;
 
@@ -235,6 +240,7 @@ impl<L> HttpTransportClientBuilder<L> {
 				connector.set_keepalive(keep_alive_duration);
 				connector.set_keepalive_interval(keep_alive_interval);
 				connector.set_keepalive_retries(keep_alive_retries);
+				connector.set_connect_timeout(connect_timeout);
 				HttpBackend::Http(Client::builder(TokioExecutor::new()).build(connector))
 			}
 			#[cfg(feature = "tls")]
@@ -251,6 +257,7 @@ impl<L> HttpTransportClientBuilder<L> {
 				http_conn.set_keepalive(keep_alive_duration);
 				http_conn.set_keepalive_interval(keep_alive_interval);
 				http_conn.set_keepalive_retries(keep_alive_retries);
+				http_conn.set_connect_timeout(connect_timeout);
 
 				let https_conn = match certificate_store {
 					CertificateStore::Native => {